Move named_command
